Rename components and add comments
